+++ Backend/src/main/java/com/pathsnap/Backend/S3/Controller/S3Controller.java
@@ -1,18 +1,13 @@
 package com.pathsnap.Backend.S3.Controller;
 
-import com.pathsnap.Backend.Exception.S3NotFoundException;
 import com.pathsnap.Backend.S3.Dto.Res.S3ListResDto;
-import com.pathsnap.Backend.S3.Dto.Req.S3UpdateReqDto;
 import com.pathsnap.Backend.S3.Dto.Req.S3UploadReqDto;
-import com.pathsnap.Backend.S3.Dto.Res.S3ResDto;
 import com.pathsnap.Backend.S3.Service.S3DeleteService;
-import com.pathsnap.Backend.S3.Service.S3Service;
+import com.pathsnap.Backend.S3.Service.S3UploadService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.IOException;
 import java.util.List;
 
 @RestController
@@ -20,14 +15,14 @@ import java.util.List;
 @RequestMapping("/images")
 public class S3Controller {
 
-    private final S3Service imageService;
+    private final S3UploadService s3UploadService;
 
     private final S3DeleteService deleteService;
 
     //S3 이미지 업로드
     @PostMapping
     public ResponseEntity<List<S3ListResDto>> uploadImages(@ModelAttribute S3UploadReqDto imageReqDTO) {
-        return ResponseEntity.ok(imageService.uploadImages(imageReqDTO));
+        return ResponseEntity.ok(s3UploadService.uploadImages(imageReqDTO));
     }
